Separate regex-based recordings from exact matches
Regex-based recordings use a different key (“uri-regex”) from exact match recordings. When recordings are added to a cassette, those that match on a regex are stored in a separate array instead of the VCRRequestKey-based dictionary.

I’m keeping the regex in a separate property on VCRRecording, too, and it’s stored as an NSRegularExpression instance instead of NSString so we don’t have to rebuild the instances on every network request.