Pass auth challenges through to NSURLProtocol client
This allows VCRURLConnection to record HTTPS responses from a server that uses a self-signed certificate. A self-signed certificate is presented to the NSURLConnection delegate as an authentication challenge. If the application normally allows this (by implementing either -connection:canAuthenticateAgainstProtectionSpace: and -connection:didReceiveAuthenticationChallenge: OR just -connection:willSendRequestForAuthenticationChallenge:), the NSURLProtocol must pass this through or the real request can’t be recorded.